+++ changelog.md
@@ -7,7 +7,6 @@
     The amount of output depends on the ore, so e.g. Lapis ores give 14 lapis (Also copper ore output has been increased)  
     Some might also be better than using fortune
   * Fixed a glitch where you could double the output from Blast furnaces just by using <REDACTED>
-* Some features can no longer be disabled (as had no effect)
 * Keego
   * Heavily reduced efficiency and attack speed of keego tools, but increased momentum strength
     * Efficiency: 6.5 -> 3.5
@@ -17,7 +16,11 @@
   * Movement Momentum now stacks faster the more pieces you have instead of having a cap on the effect level
   * Increased movement momentum bonus speed (+5% -> +6% per level)
   * Armor speed penalty has been increased (10% -> 20%)
+* Recall potion is now made from a Withering Regeneration potion and echo shard
+  * Regeneration potion + Wither Rose = Withering Regenerating
+  * Withering Regeneration + Echo shard = Recall potion
 * Reduced burst of arrows number of arrows (8 -> 6)
+* Some features can no longer be disabled (as had no effect)
 * Added a starting season config option
 
 ## 1.6.8